+++ changelog.md
@@ -26,6 +26,9 @@
 * Modules
   * Fixed a bug where a module would not be started without an explicit definition for that module (#388 & #400)
 
+* All Views
+  * Allow `modelEvents` and `collectionEvents` to be a function that returns a hash
+
 ### v1.0.0-rc2 [view commit logs](https://github.com/marionettejs/backbone.marionette/compare/v1.0.0-rc1...v1.0.0-rc2)
 
 * CollectionView / CompositeView